zoom in zoom out in mermaid graph. #774.
